refactor(dpmodel): group NeighborGraph into a subpackage
Move the flat utils modules (neighbor_graph.py, segment.py, edge_derivatives.py)
into deepmd/dpmodel/utils/neighbor_graph/ as a cohesive subpackage:
  graph.py (contract + node-validity + edge padding), builder.py
  (neighbor_graph_from_extended, build_neighbor_graph), segment.py, derivatives.py.
Public API re-exported from the package __init__ and from utils/__init__, so
import paths (deepmd.dpmodel.utils.*) are unchanged.
