+++ deepmd/dpmodel/utils/neighbor_graph/segment.py
